fix over-agressive kramdown config
The `parse_block_html` config causes any text indented at four spaces
to be treated as a code block, which messes up various parts of the
site. I set this so that the table of contents would work as desired
on the context post. This commit reverts the config, and explicitly
sets the `markdown` attribute in the wrapper div of the toc, which
results int he desired behavior.